[#2] YDBAIM CMake Installer
Install instructions described in README. Standard `cmake .. && make &&
sudo make install`. Tested on yottadb-base docker image.

Note that we should technically test if we are above a certain version,
but this is somewhat difficult to do w/o a lot of scripring in CMake; so
we will keep that as a perhaps future enhancement, as many of the
plugins don't do it. Technically, we need to be at r132 or above. I
would personally vote to postpone this until the YottaDB CMake
capability can provide a lot more features than what we have now. Octo
has a bash script that does the job.