Add testing system to CI.
.gitlab-ci.yml now has a new `make-ubuntu` job that uses the
`yottadb/yottadb-base:latest-master` as the base, installs YDBAIM, and
then runs the Unit Tests.

`tests/test.sh` enhanced to tee the M-Unit output, and then grep for
`[FAIL]`. If `[FAIL]` is found, the test that caused the failure is
printed in the same grep (using the -B1 flag), and we exit with an
error. Some gymnastics required because grep's return status is the
opposite of what we want.

Lastly, the failing test in `tests/_YDBAIMTEST.m` has been temporarily
removed. It's tracked in
https://gitlab.com/YottaDB/Util/YDBAIM/-/issues/26. Now pipeline
succeeds.

Other misc changes:

- Change gpgkey-verify to commit-verify.