Check that YottaDB is installed with UTF-8 support before attempting to install YDBAIM in UTF-8 mode
This commit changes install.sh to make sure that YottaDB was installed with UTF-8 support before attempting to install YDBAIM in UTF-8 mode. Previously, the install script would do this, causing a `ydbinstall --aim` command to attempt to install YDBAIM in UTF-8 mode which would fail and produce an error message.