Increase minimum CMake version to 3.0.0
This commit addresses CMake deprecation warnings by raising the minimum CMake version to 3.0.0. The CMake warnings say that compatibility with versions < 2.8.12 will be removed in a future CMake version. Since the current minimum for the YDB repo is 3.0.0, it makes sense to increase the minimum CMake version to 3.0.0 to be consistent with the YDB repo.