Fix UTF-8 check to check if $ydb_dist exists before checking for UTF-8 in pkg-config
This commit fixes a bug in YDBAIM!24. The code was intended to check for UTF-8 support in $ydb_dist if defined or in pkg-config if $ydb_dist is undefined. However, it checks if either $ydb_dist or pkg-config points to a YottaDB version with UTF-8 support. This causes YDBAIM to attempt a UTF-8 build that will fail due to lack of UTF-8 support if $ydb_dist points to a YottaDB version without UTF-8 support and pkg-config points to a version with UTF-8 support. The fix is to check that $ydb_dist is undefined before checking if pkg-config points to a YottaDB version with UTF-8 support.