[#42] [#44] [#45] Speed up initial cross referencing; address bugs & misfeatures in error handler
For the intial metadata calculation of existing data, look only for those subscripts which are in the range(s) specified for cross referencing. The previous code checked all subscripts and skipped past those not in range to be indexed. Also a very minor performance improvement by replacing all checks for equality with the empty string (`""`) with '$ZLENGTH().

Return to top level in the event of an error. Fix error handler:

- to remove extraneous messages;
- for top level invocation, to not report undefined tlevel; and
- make TROLLBACK conditional on `tlevel` being defined.