Remove %YDBAIMTEST trap unwinder; reenable tests
The YDBAIM code finally has an error trap that unwinds properly from
deep levels of its stack. As a result, the unwinder in the error trap is
no longer needed and is now removed. This in itself will be a good test
of the error trap unwinding properly.

In addition bash tests INVPNUMSEP and CANTADDSTAT are now reenabled.