[#52] XREFDATA() should not create metadata for nonexistent nodes when the last subscript is a constant, and there are other nodes at that level
When the last subscript is a constant, XREFDATA() does not create metadata for nonexistent application global nodes at the lowest level when there are other nodes at that level with non-matching subscripts. Fixed by adding $DATA() checks for nodes at that level when cross referencing. The code originally worked correctly, but was broken when an attempt to make the [#51] behavior the default, followed by backing out those changes.