[#26] [#51] Maintain metadata with KILLs of trees; implement type=1 metatadata
This commit enhances YDBAIM so that KILLs of trees maintain metadata. It achieves this by using KILL triggers that recursively perform depth first traversals of trees. Each tree traversal performs a ZKILL of the root node after traversing the tree, and any metadata at that level is maintained by the ZKILL trigger. All KILL triggers use this technique. The extra triggers this causes have suffixes from the sequence `0123456789ABCDEFGHIJLMNOPQRTUVWX`, since K, S and Z are used as suffixes for KILL, SET, and ZKILL triggers respectively.

Note: There is an alternative technique which would examine the triggers on a global variable, and simply KILL trees below the deepest level for which there is a trigger. This has the overhead of obtaining and analyzing the triggers on a global variable, a cost and benefit over recursive traversal that would be well amortized for KILLing a large tree. As it is anticipated that the majority of KILLs used for globals for which AIM maintains metadata are small trees, the depth first traversal technique was chosen. To implement this in the future, the simplest way would be to create new types that match existing types but have different KILL triggers.

Triggers for `type=1` global variables are similar to the default (`type=""`) global variables, except that the SET and ZKILL triggers maintain alternate metadata nodes. Note that if a global variable node is deleted, with alternate metadata node created because it has siblings.

As minor optimizations, $ZTRIGGER() with a trigger names is used for removing triggers, and piece numbers are determined at code generation time, not execution time.

There are two new error messages (U235 & U236), and one error message with different text (U253). A few comments are improved.

Test changes:
- Test added for `type=1` subscripts
- Line number changes in `bash_test_output.ref`
- `tsubs1` test corrected to KILL ^ORD