[#61] Fix rare %YDB-E-GVUNDEF error from UNXREFDATA^%YDBAIM
With concurrent deletes of a ^%ydbAIMD* metadata global variable, there was a race condition between testing for the existence of such a global variable and its deletion, leading to rare GVUNDEF errors. An earlier fix that used M Locks still had a potential race condition. Use of TP prevents it.

Enhance tests to check for ^%ydbAIMDxref deletion.