[#66] Concurrent cross-referencing passes in the pipeline on a 16cpu machine
The issue was that a TSTART needed to protect the zero initial value of a local variable when counting the total number of application nodes in the (11) node of the cross reference global. It does with this commit.

A minor correction to the README is included, that the tests need to be run in the `tests` subdirectory.