[#69] Fix tbash test: subtest GVSUBOFLOW0 or GVSUBOFLOW1
This test fails because since the implementation of #42, it is possible
that this test (which intentionally produces an error) to produce two
lines for the error message, since #42 made it so that two concurrent
processes perform the initial metadata computation.

The solution is to pipe the output to `uniq`; The status line was
adjusted to get the status of the `yottadb` run, not the status of the
`uniq` command.