[#68] [#70] Fix interrupt issues; undefined variable errors; indexing a big global producing an index with missing data + Tests
Fix undefined variable errors in XREFDATAQUIT^%YDBAIM.

Process interrupts as follows:

  - For USR1, XECUTE the $ZINTERRUPT of the parent and continue.
  - For USR2, terminate %YDBAIM and rethrow the interrupt.

Fix indexing a big global producing an index with missing data:

With two processes cooperating on the intial scan of a global variable, one scanning up from earlier collating subscripts and the other scanning down from later collating subscripts, the main %YDBAIM process is designed to periodically review the progress of the proceses, and stop them when the process scanning up is scanning the global variable tree past the location of the process scanning down. This had a subtle error in that if the two processes had the same subscript value at a certain level, it would consider the subscript scanned at that level, whereas the scanning is complete only when the subscript of the first process collates after the subscript of the second process. Since the parent process would only scan intermittently, and since it was quite likely that the the time the parent process delivered the signals to terminate the scanning processes, in most cases the scan would complete, and bug only occurred every tens to hundreds of runs. With the fix, several thousand runs have passed.

Two other fixes made at the same time:

- The code in the parent process checking the subcripts of the child processes needed the check for the existence of a subscript and accessing its value to be in a transaction (a $GET() would not work as a reasonable default cannot be provided).
- In the event of a normal completion. XREFDATA() now restores the value of $INTERRUPT. The older code restored it in the event of termination with a USR2, but not for a normal completion.

Tests from @shabiel:

 - tsigusr1 tests AIM behavior with SIGUSR1
 - tsigusr2 tests AIM behavior with SIGUSR2
 - tintrestore test to test that interrupts are restored.

Also:

- Change line numbers to account for code changes and to allow bash test to pass
- Fix copyright year check pipeline failure