[#72] XREFDATA^%YDBAIM continues cross referencing after first value of lowest level subscript if xsub has range (:) syntax with empty range-end
When cross referencing the lowest level subscript, if a range is specified without an end, e.g., `2:`, the logic for processes scanning nodes in ascending order would have terminated the loop if the actual subscript collated after the empty string. This would cause it to terminate after the first value. The logic for higher level subscripts was correct, and the fix was to apply that logic to the lowest level subscript.

The bug was intermittent, because sometimes the process scanning nodes in the reverse direction would cross reference those nodes. But sometimes, the process scanning in ascending order would conclude that that it was done with its cross referencing, and terminate the process scanning in the descending order. In this latter case, there would be missing cross references.

%YDBAIMTEST includes a regression test for the fix.

Responses to review comments are included.