[#73] Fix random pipeline failures
Previously, there were random, rare assert failures in the aim73 test discussed in https://gitlab.com/YottaDB/Util/YDBAIM/-/merge_requests/70#note_1643645543 where the test would set a node and expect the trigger to increase the count of the number of ndoes by 1. However, every once in a while, the node that was set would exist already. This would cause an assert failure as the count would not increase. The change ensures that the new node has subscripts that are guaranteed not to already exist. The initial nodes have subscripts that are 3-character random alphanumeric strings. The new nodes set have 4, 5, and 6 character long random subscripts.

Debugging code is removed from this commit.