[#74] Literal string subscripts that include "/" work
This fixes the problem documented at https://gitlab.com/YottaDB/DBMS/YDBOcto/-/issues/1034#note_1804235992

Literal string subscripts that include "/" now work. The problem was that "/" characters in literal subscript strings were interfering with the "/" characters used as option separators in templates for different application subscript types (AIMTYPEs). The solution was to move the option selection prior to the substitution to create the trigger code from templates. A side effect of this change is that strings during substitution are marginally smaller.