[#5] Metadata for subscripts
This implements the functionality described in the Release Notes of #5, including removal of duplicated subscripts on numerical scans. That description is not duplicated here.

The code that cross references subscripts broadly follows the structure of code that cross references pieces of nodes. Note that that it does not implement special cross referencing for subscripts of Fileman schema globals, as it was determined that normal cross-referencing for such subscripts suffices. The implementation does not preclude adding special cross referencing in the future, should the need arise.

The code for *DATA() functions is unchanged, except that the routine is refactored to combine code common to *DATA() and *SUB().

New labels are

  - LSXREFSUB() closely mirrors LSXREFDATA(), and invokes lsxref() to list the cross references.
  - UNXREFSUB() follows the structure of UNXREFDATA(), and invokes unxref() to remove the cross references.
  - XREFSUB() follows the structure of XREFDATA() but differs in that it needs to handle subscripts. The call to $ZYSUFFIX() is salted to ensure that the hash-generated cross reference names differ; without the salt, they would be the same.
  - initxref is initialization code common to XREFDATA() and XREFSUB().
  - mkindxrefsub creates indirection strings to be used in the initial cross-referencing of subscripts.

The following labels are renamed since they now handle both *DATA() and *SUB() invocations

  - lsxrefdata() is now lsxref().
  - unxrefdata() is now unxref().

Management of the two parallel processes that perform the initial scan of a global variable has been refactored into common code.

A significant difference between the implementation of triggers for pieces of nodes vs. subscripts is that whereas the pieces to be cross referenced are iterated over in a FOR loop, each subscript has its own code. This is because (a) $[Z]PIECE() is believed to be more efficient than $QSUBSCRIPT(); and (b) while global variables can have large numbers of pieces, global variables have much smaller numbers of subscripts. To simplify the trigger templates for subscripts, a new function expline() creates multi-line trigger templates from single-line trigger templates. The function exptempl() is modified to invoke expline(). This change also resulted in a minor imrovement to trigger template consistency in that single-line triggers no longer need to start with a space. There are new triggers for managing subscripts.

Other changes
  - During the initial scan, the processes scanning in either direction update their status every 10msec
  - New error messages were added, and a couple of error messages changed.
  - The maximum number of application subscripts AIM can handle is 29 not 31, since AIM has an overhead of two subscripts. Checks on the number of subscripts were changed from 31 to 29.

The Munit tests were updated to add tests for subscripts. Each existing test for *DATA() functionality was reviewed as to whether a similar test was meaningful for *SUB() functionality. Except for those small number of tests, existing tests were enhanced in one two ways:

  - The most common way was to create code to test *SUB() functions below the code to test *DATA() functions.
  - Where it seemed appropriate, the additional code was interleaved into the existing code.

Other:

- Split speed tests
- Only do the database expansion for pipeline test
- Make Octo test pass (@shabiel was able to get one to fail)
- Minor non-material clean up changes to comments, and occasionally code.
- Tests enhanced for *SUB() functions.
- Responses to multiple rounds of review comments.