[#77] Additional type to enable Octo key fixing optimization
With a `type=2` parameter, XREFDATA^%YDBAIM() expects the `force` parameter to be a string that specifies a labelref of a transformation function, e.g., `"$$ABC^DEF()"`. Triggers for cross referenced nodes will use the value returned by the transformation function as the value to cross reference. When the function is called at runtime by the trigger, the first parameter is the actual node or piece value, e.g., `$$ABC^DEF("2024-02-21T13:31:48.05098021+07:00")` would be the actual call using the example of a timestamp at [Support timestamp values of the format 2024-02-21T13:31:48.05098021+07:00](YottaDB/DBMS/YDBOcto#1050). If the function requires additional parameters, they can be specified as comma separated values for the second and subsequent parameters, e.g., `"$$ABC^DEF(,1,""two"")"`. As local variables cannot be passed to triggers, these additional parameters can only be constants, global variable references, or function calls whose parameters are constants, global variables, or function calls. Application code that needs to pass local variable values to the transformation function should use [$ZTWORMHOLE](https://docs.yottadb.com/ProgrammersGuide/isv.html#ztwormhole).

There are additional tests, for the illegal combinations of `force` and `type`, as well as tests for code that was not previously tested. There are also minor bug fixes and cleanup, as well as a small test change necessitated by YottaDB r2.00, whose default key size is 1,019 bytes, the maximum.

The commit includes responses to review comments.