[YottaDB/DBMS/YDBOcto#1083] Run TRANS2BIG test ONLY on pipeline
This commit modifies the previous commit
6c5849bba03d6decc932688ab6d1db7604ebfc75 that implemented this test.

This test now runs only on the pipeline because it's outside of the main test suite
that is used for development; so having it run during development is not ideal,
as it interrupts developement workflow (e.g. runs every time since it's
outside the test suite, and has verbose output).