[#83] Implement a way to find out the version of AIM being used
This implements the functionality described in #83.

THe version number on the comment the %YDBAIM label is now the minor
version number. Comments on the XREFDATA() and XREFSUB() labels are the
major version numbers for the data and subscript schemas respectively. A
new VERSION label has the code to report the version.

There is a new version automated test, and tbash test line numbers have
been adjusted.

`CMakeLists.txt` has been modified to add a manifest file so that the
YottaDB GUI can now report AIM's version number.