Enhance assert() in _YDBAIMTEST.m to create a jobexam file for helping with debugging rare failures
We found a test that failed on a pipeline with this error:

https://gitlab.com/YottaDB/Util/YDBAIM/-/pipelines/2225525601

```
trange4^%YDBAIMTEST - indexing subscripts with an open range - Error from trange4+5^%YDBAIMTEST: do assert($data(@xref@(0,"abcd")))
-------------------------------------------------------------- [FAIL]
```

The test is very simple, and running it on multiple hosts about 150000
times failed to reproduce the error.

For now, we are enhancing the test framework to do jobexam's when an
assert fails. This will dump local variables as well as the stack for
later examination. This will be part of the artifacts.

Sample pipeline with an artifact on failure: https://gitlab.com/shabiel/YDBAIM/-/jobs/12532788746